+++ packages/testing/src/api/routes/oidc.franceconnect.gouv.fr/select.page.tsx
@@ -1,7 +1,7 @@
 //
 
 import { type Citizen } from "#src/api/data/franceconnect";
-import { type FranceConnectUserInfoResponse } from "@gouvfr-lasuite/proconnect.identite/types";
+import { type FranceConnectUserInfoResponse } from "@proconnect-gouv/proconnect.identite/types";
 
 //